fixes windows OpenMP error related to unsigned variables
when building with OpenMP on windows there is an error related to
unsigned variables:

'variable': index variable in OpenMP 'for' statement must have signed
integral type

So hereby added i8, i16, i32 and i64 types for use with OpenMP.

+++ src/par2creator.cpp
@@ -929,7 +929,7 @@ bool Par2Creator::ProcessData(u64 blockoffset, size_t blocklength)
 
     // For each output block
     #pragma omp parallel for
-    for (u32 outputblock=0; outputblock<recoveryblockcount; outputblock++)
+    for (i64 outputblock=0; outputblock<recoveryblockcount; outputblock++)
     {
       // Select the appropriate part of the output buffer
       void *outbuf = &((u8*)outputbuffer)[chunksize * outputblock];
